altera o tipo de fps de double para int na configuração do vídeo do Vimeo

+++ lib/src/model/vimeo_video_config.dart
@@ -94,7 +94,7 @@ class VimeoVideoFile {
   int? height;
   String? link;
   String? createdTime;
-  double? fps;
+  int? fps;
   int? size;
   String? md5;
   String? publicName;
